Works how I want it to (TEST)
(TEST FEATURES ADDED-NOT FINAL)
Can sort through an array, and sort between equal, and not equal. This
means that it can detect weather or not the distance needs to go into
the  corrector or not. I don’t think it is concurrent yet, still
sequential as it takes right input first. That is the next big step in
this. Once I am able to tell how it receives input I can better
implement the concurrency.